requirements.txt
coreapi==2.3.3
Django==3.2.9
django-extensions==3.1.5
django-filter==21.1
django-rest-framework==0.1.0
djangorestframework==3.12.4
ipython==7.29.0
Werkzeug==2.0.2
celery==5.2.1
requests~=2.26.0
redis==3.5.3
